Add CustomInput component
CustomInput component accepts props for title, value, handleChangeText, otherStyles, keyboardType, placeholder, and type. It includes a TextInput field with dynamic styles and functionality based on the provided props. For password input, it includes a toggle button for showing and hiding the password. The state of the password visibility is managed using a local state variable, showPassword.