Implement Home component
The Home displays a list of posts in a FlatList. The posts are fetched using the useAppwrite custom hook with the getAllPosts function. The component includes a refresh control for the FlatList that allows the user to refresh the posts. It also includes a ListHeaderComponent that displays a welcome message, a SearchInput component, and a Trending component. If there are no posts, it displays an EmptyState component.