Add searchPosts function for Appwrite database
This commit introduces a new function, searchPosts, which allows for searching posts in the Appwrite database based on a given query. The function uses the Appwrite SDK to list documents in a specified video collection that match the search query. The search operation is performed on the 'title' field of the documents in the video collection. The search query is case-insensitive and supports partial matches. The function returns a promise that resolves to an array of post documents that match the search query. If the function fails to retrieve the posts, the promise is rejected with an error.