Add Search component for displaying search results
This commit introduces a new component, Search, which displays the search results from the Appwrite database. The component uses the 'expo-router' package to get the search query from the URL parameters. It then uses the 'useAppwrite' hook to call the 'searchPosts' function with the query. The search results are displayed in a FlatList, with each item rendered using the VideoCard component. If there are no search results, an EmptyState component is displayed. The search query can be updated using the SearchInput component, which is included in the ListHeaderComponent of the FlatList.